+++ apis/nucleus/src/components/listbox/__tests__/list-box-inline.spec.jsx
@@ -96,7 +96,7 @@ describe('<ListboxInline />', () => {
         [require.resolve('../../../hooks/useSessionModel'), () => useSessionModel],
         [require.resolve('../../../hooks/useLayout'), () => () => [layout]],
         [require.resolve('../../ActionsToolbar'), () => ActionsToolbar],
-        [require.resolve('../ListBox'), () => <div className="TheListBox" />],
+        [require.resolve('../ListBox'), () => <div className="theListBox" />],
         [require.resolve('../ListBoxSearch'), () => ListBoxSearch],
         [
           require.resolve('../listbox-keyboard-navigation'),
@@ -118,6 +118,7 @@ describe('<ListboxInline />', () => {
       on: sandbox.stub().callsFake((event, func) => (eventTriggered) => {
         if (event === eventTriggered) func();
       }),
+      off: sandbox.stub(),
     };
     useSessionModel.returns([sessionModel]);
     useObjectSelections.returns([selections]);
@@ -133,6 +134,7 @@ describe('<ListboxInline />', () => {
       sessionModel: undefined,
       selectionsApi: undefined,
       update: undefined,
+      fetchStart: 'fetchStart',
     };
 
     theme.spacing.returns('padding');
@@ -171,6 +173,7 @@ describe('<ListboxInline />', () => {
 
   afterEach(() => {
     sandbox.reset();
+    renderer.unmount();
   });
 
   after(() => {
@@ -213,6 +216,11 @@ describe('<ListboxInline />', () => {
       expect(showSearchButtons).to.have.length(1);
       expect(getListboxInlineKeyboardNavigation).calledOnce;
       expect(renderer.toJSON().props.onKeyDown).to.equal('keyboard-navigation');
+
+      expect(selections.on).calledTwice;
+      expect(selections.on.args[0][0]).to.equal('deactivated');
+      expect(selections.on.args[1][0]).to.equal('activated');
+      expect(selections.off).not.called;
     });
 
     it('should render without toolbar', async () => {