+++ apis/nucleus/src/components/listbox/__tests__/list-box.spec.jsx
@@ -13,10 +13,15 @@ describe('<Listbox />', () => {
   let pages;
   let selectDisabled;
   let FixedSizeList;
+  let fetchStart;
+  let useCallbackStub;
+  let setTimeoutStub;
   let useSelectionsInteractions;
 
   before(() => {
     sandbox = sinon.createSandbox({ useFakeTimers: true });
+
+    setTimeoutStub = sandbox.stub();
 
     global.document = 'document';
 
@@ -31,6 +36,9 @@ describe('<Listbox />', () => {
 
     useSelectionsInteractions = sandbox.stub();
 
+    fetchStart = sandbox.stub();
+    useCallbackStub = sandbox.stub();
+
     FixedSizeList = sandbox.stub().callsFake((funcArgs) => {
       const { children } = funcArgs;
       const RowOrColumn = children;
@@ -41,6 +49,13 @@ describe('<Listbox />', () => {
       [
         [require.resolve('react-window'), () => ({ FixedSizeList })],
         [require.resolve('../../../hooks/useLayout'), () => () => [layout]],
+        [
+          require.resolve('react'),
+          () => ({
+            ...React,
+            useCallback: useCallbackStub,
+          }),
+        ],
         [require.resolve('../useSelectionsInteractions'), () => useSelectionsInteractions],
         [
           require.resolve('react-window-infinite-loader'),
@@ -64,6 +79,8 @@ describe('<Listbox />', () => {
   beforeEach(() => {
     pages = [{ qArea: { qTop: 1, qHeight: 100 } }];
 
+    global.window = { setTimeout: setTimeoutStub };
+
     selectDisabled = () => false;
 
     useSelectionsInteractions.returns({
@@ -86,9 +103,9 @@ describe('<Listbox />', () => {
       width: 100,
       listLayout: 'vertical',
       update: sandbox.stub(),
-      rangeSelect: false,
       checkboxes: false,
       selectDisabled,
+      fetchStart,
     };
   });
 
@@ -102,18 +119,20 @@ describe('<Listbox />', () => {
 
   describe('Check rendering with different options', () => {
     before(() => {
-      render = async () => {
+      render = async (overrides = {}) => {
+        const mergedArgs = { ...args, ...overrides };
         await act(async () => {
           renderer = create(
             <ListBox
-              selections={args.selections}
-              direction={args.direction}
-              height={args.height}
-              width={args.width}
-              rangeSelect={args.rangeSelect}
-              listLayout={args.listLayout}
-              update={args.update}
-              checkboxes={args.checkboxes}
+              selections={mergedArgs.selections}
+              direction={mergedArgs.direction}
+              height={mergedArgs.height}
+              width={mergedArgs.width}
+              listLayout={mergedArgs.listLayout}
+              update={mergedArgs.update}
+              checkboxes={mergedArgs.checkboxes}
+              selectDisabled={mergedArgs.selectDisabled}
+              fetchStart={mergedArgs.fetchStart}
             />
           );
         });
@@ -126,7 +145,6 @@ describe('<Listbox />', () => {
     });
 
     it('should render and call stuff', async () => {
-      args.rangeSelect = false;
       await render();
 
       // check rendering
@@ -143,7 +161,6 @@ describe('<Listbox />', () => {
         layout,
         selections,
         pages: [],
-        rangeSelect: false,
         doc: 'document',
       });
       expect(typeof useSelectionsInteractions.args[1][0].selectDisabled).to.equal('function');
@@ -161,17 +178,35 @@ describe('<Listbox />', () => {
       expect(itemData.onClick).to.equal(undefined);
     });
 
-    it('should call useSelectionsInteractions with rangeSelect true', async () => {
-      args.rangeSelect = true;
+    it('should call useSelectionsInteractions', async () => {
       await render();
       expect(useSelectionsInteractions.args[useSelectionsInteractions.callCount - 1][0]).to.containSubset({
         checkboxes: false,
         layout,
         selections,
         pages: [],
-        rangeSelect: true,
         doc: 'document',
       });
+    });
+
+    it('should not call fetchStart unless fetching data', async () => {
+      sandbox
+        .stub(React, 'useRef')
+        .onFirstCall()
+        .returns({
+          loaderRef: {
+            current: {
+              _listRef: { state: { isScrolling: false } },
+            },
+          },
+        })
+        .callsFake((inp) => ({ current: inp }));
+
+      await render();
+      expect(fetchStart).not.called;
+      expect(setTimeoutStub).not.called;
+      const loadMoreItems = useCallbackStub.args[1][0];
+      expect(loadMoreItems).to.be.a('function');
     });
 
     it('should call with checkboxes true', async () => {
@@ -182,7 +217,6 @@ describe('<Listbox />', () => {
         layout,
         selections,
         pages: [],
-        rangeSelect: false,
         doc: 'document',
       });
       const { itemData } = FixedSizeList.args[FixedSizeList.callCount - 1][0];