+++ apis/nucleus/src/components/listbox/__tests__/use-selections-interactions.spec.jsx
@@ -65,9 +65,7 @@ describe('use-listbox-interactions', () => {
           <TestHook
             ref={ref}
             hook={useSelectionsInteractions}
-            hookProps={[
-              { layout, selections, rangeSelect: false, pages, selectDisabled, doc: global.document, ...overrides },
-            ]}
+            hookProps={[{ layout, selections, pages, selectDisabled, doc: global.document, ...overrides }]}
           />
         );
       });
@@ -85,15 +83,8 @@ describe('use-listbox-interactions', () => {
 
   describe('it should behave without range select', () => {
     describe('should return expected listeners', () => {
-      it('Without range', async () => {
-        await render();
-        const arg0 = ref.current.result;
-        expect(Object.keys(arg0).sort()).to.deep.equal(['instantPages', 'interactionEvents', 'select']);
-        expect(arg0.instantPages).to.deep.equal([]);
-        expect(Object.keys(arg0.interactionEvents).sort()).to.deep.equal(['onMouseDown', 'onMouseUp']);
-      });
       it('With range', async () => {
-        await render({ rangeSelect: true });
+        await render();
         const arg0 = ref.current.result;
         expect(Object.keys(arg0).sort()).to.deep.equal(['instantPages', 'interactionEvents', 'select']);
         expect(arg0.instantPages).to.deep.equal([]);
@@ -195,7 +186,7 @@ describe('use-listbox-interactions', () => {
 
   describe('it should behave with range select', () => {
     it('should return expected stuff', async () => {
-      await render({ rangeSelect: true });
+      await render();
       const arg0 = ref.current.result;
       expect(Object.keys(arg0)).to.deep.equal(['instantPages', 'interactionEvents', 'select']);
       expect(arg0.instantPages).to.deep.equal([]);
@@ -205,7 +196,7 @@ describe('use-listbox-interactions', () => {
     it('should select a range (in theory)', async () => {
       getElemNumbersFromPages.returns([24, 25, 26, 27, 28, 29, 30, 31]);
 
-      await render({ rangeSelect: true });
+      await render();
 
       expect(applySelectionsOnPages.callCount).to.equal(0);
 
@@ -278,7 +269,43 @@ describe('use-listbox-interactions', () => {
           },
         });
       });
-      expect(applySelectionsOnPages.args[startCallCount + 1]).to.deep.equal([[], [24], false]);
+      expect(applySelectionsOnPages.args[startCallCount + 1]).to.deep.equal([[], [], false]);
+    });
+
+    it('Ctrl or cmd button with click should result in single select behaviour', async () => {
+      await render({ checkboxes: true });
+      const preventDefault = sandbox.stub();
+      const focus = sandbox.stub();
+      await act(() => {
+        ref.current.result.interactionEvents.onClick({
+          currentTarget: {
+            focus,
+            getAttribute: sandbox.stub().withArgs('data-n').returns(24),
+          },
+          ctrlKey: true,
+          preventDefault,
+        });
+      });
+      expect(focus.callCount).to.equal(1);
+      expect(preventDefault.callCount).to.equal(1);
+    });
+
+    it('Ctrl or cmd button with mousedown should result in single select behaviour', async () => {
+      await render({ checkboxes: false });
+      const preventDefault = sandbox.stub();
+      const focus = sandbox.stub();
+      await act(() => {
+        ref.current.result.interactionEvents.onMouseDown({
+          currentTarget: {
+            focus,
+            getAttribute: sandbox.stub().withArgs('data-n').returns(24),
+          },
+          ctrlKey: true,
+          preventDefault,
+        });
+      });
+      expect(focus.callCount).to.equal(1);
+      expect(preventDefault.callCount).to.equal(1);
     });
   });
 });